+++ tests/test_order_group_scope_relative.py
@@ -6,10 +6,7 @@ import pytest
 
 import pytest_order
 
-try:
-    from tests.utils import write_test, assert_test_order
-except ImportError:
-    from utils import write_test, assert_test_order
+from tests.utils import write_test, assert_test_order
 
 
 @pytest.fixture(scope="module")