+++ tests/test_order_scope.py
@@ -5,10 +5,7 @@ import shutil
 import pytest
 
 import pytest_order
-try:
-    from tests.utils import write_test, assert_test_order
-except ImportError:
-    from utils import write_test, assert_test_order
+from tests.utils import write_test, assert_test_order
 
 
 @pytest.fixture(scope="module")