+++ tests/test_relative_ordering.py
@@ -5,11 +5,7 @@ import shutil
 import pytest
 
 import pytest_order
-
-try:
-    from tests.utils import write_test, assert_test_order
-except ImportError:
-    from utils import write_test, assert_test_order
+from tests.utils import write_test, assert_test_order
 
 
 def test_relative(item_names_for):